kokanystream: Merge into a single script
There have been hardware changes the code had to be adapted to.

We switched to two identical USB webcams, so we need to identify them by
their alternative names. Since it doesn't really matter which camera is
the front or back because kokanyctl is written so well (/s) that both
work just fine, we just merged the two scripts into one and now we
simply use an index to select the camera&port combo we need.